+++ src/components/shell/index.ts
@@ -1,3 +1,4 @@
 export { AppShell } from './AppShell'
 export { CloseConfirmDialog } from './CloseConfirmDialog'
 export { QuitConfirmDialog } from './QuitConfirmDialog'
+export { UpdaterFailureDialog } from './UpdaterFailureDialog'